Move time utils to timestuff
